relates to #1839: move the QueryConfig to the common project
