+++ apis/sgv2-docsapi/src/main/java/io/stargate/sgv2/docsapi/service/query/executor/QueryExecutor.java
@@ -29,9 +29,9 @@ import io.smallrye.mutiny.converters.multi.MultiRx3Converters;
 import io.stargate.bridge.proto.QueryOuterClass;
 import io.stargate.bridge.proto.StargateBridge;
 import io.stargate.sgv2.api.common.StargateRequestInfo;
+import io.stargate.sgv2.api.common.config.QueriesConfig;
 import io.stargate.sgv2.docsapi.api.common.properties.document.DocumentProperties;
 import io.stargate.sgv2.docsapi.api.common.properties.document.DocumentTableProperties;
-import io.stargate.sgv2.docsapi.config.QueriesConfig;
 import io.stargate.sgv2.docsapi.service.ExecutionContext;
 import io.stargate.sgv2.docsapi.service.common.model.RowWrapper;
 import io.stargate.sgv2.docsapi.service.query.model.ImmutableRawDocument;
