+++ apis/sgv2-docsapi/src/main/java/io/stargate/sgv2/docsapi/service/schema/query/JsonSchemaQueryProvider.java
@@ -17,8 +17,8 @@
 package io.stargate.sgv2.docsapi.service.schema.query;
 
 import io.stargate.bridge.proto.QueryOuterClass;
+import io.stargate.sgv2.api.common.config.QueriesConfig;
 import io.stargate.sgv2.api.common.cql.builder.QueryBuilder;
-import io.stargate.sgv2.docsapi.config.QueriesConfig;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
